InfluxDB binding includes prometheus remote urls
So it is easy to use InfluxDB as a Prometheus remote, we should include
the URLs preformatted for Prometheus in the binding, which will appear
in VCAP services.

This commit adds a bunch of embedded structs, some of which are optional

This allows us to use the same struct for return type but for it to have
optional fields.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>